Increase the initial number of worker threads
This patch sets the initial number of threads to the number of available
CPUs in the system. The original initial number of threads is 2.  This
is conservative because it will not utilize all available CPUs in the
beginning of applications. An example is the test case dispatch_concur
in libdispatch in Linux OS. This test expects all CPUs will be utilized
for tasks in the global concurrent queue, but libpwq only creates 2
worker threads to run tasks in the global queue. As a result, the test
fails in machines with 8 or more CPUs.