Use CPU affinity for number of available CPUs.
Currently the number of online CPUs is considered the number
of CPUs available to the process, but taskset/numactl/etc
can be used to bind the process to specific CPUs.

In cases where the process is bound to fewer than the number of
online CPUs, the number of worker threads created can be too
many and the idle threshold too high, leading to more CPU
contention and poorer performance.