Implement signal semantics expected by libdispatch
As discussed in https://bugs.swift.org/browse/SR-2655,
libdispatch expects that all non-synchronous signals
are masked on worker threads. The only signals that
should be unmasked are the ones that are delivered to
the generating thread itself. This patch uniformly
implements those semantics for all threads created
by libpwq (manager, normal and overcommit workers).