build: add support for installing from CMake
We can now install the library, manpage, and pkg-config data from CMake.
At this point, for Linux, FreeBSD, Windows, and android, cmake is
sufficient and should be treated as the canonical build system.

Bump CMake requirement to 3.4.3.  This should be old enough to be
generally available and matches what LLVM, clang, and swift support as a
minimum requirement.